requirements.txt
Flask
Flask-MySQLdb
mysqlclient
python-dotenv
Flask-Login